Add files via upload
Three simple exercises to understand the difference in logic between SPA and normal HTML pages. With SPA the browser performs a single request to the server and then the Javascript injects the different content. On the other hand, with traditional webpages, the browser needs to perform a new GET request to the server whenever the content changes.